+++ exercicio.generics/src/services/ProdutoClass.java
@@ -4,9 +4,9 @@ import java.math.BigDecimal;
 
 public abstract class ProdutoClass {
     protected String nome;
-    protected BigDecimal preco;
+    protected Float preco;
 
-    public ProdutoClass(String nome, BigDecimal preco) {
+    public ProdutoClass(String nome, Float preco) {
         this.nome = nome;
         this.preco = preco;
     }
